UHF-6038: Converted menu tree builder to new service and refactored menu updater.
