+++ helfi_features/helfi_navigation/README.md
@@ -1,0 +1,9 @@
+# HELfi navigation
+
+## Note: This module is not completed and is not ready for production (7.7.2022)
+
+Provides a module for transferring menu data across instances.
+
+Menus are converted into JSON, which is sent to Etusivu (front page) instance of Hel.fi ecosystem.
+
+JSON structure is defined in schema.json file under assets.
