+++ helfi_features/helfi_navigation/helfi_navigation.services.yml
@@ -6,14 +6,14 @@ services:
     arguments: ['helfi_navigation']
   helfi_navigation.external_menu_tree_factory:
     class: \Drupal\helfi_navigation\ExternalMenuTreeFactory
-    arguments: ['@helfi_navigation.schema_storage', '@logger.channel.helfi_navigation']
+    arguments: ['@helfi_navigation.schema_storage', '@logger.channel.helfi_navigation', '@helfi_api_base.internal_domain_resolver', '@helfi_api_base.environment_resolver', '@helfi_navigation.global_navigation_service']
   helfi_navigation.global_navigation_service:
     class: \Drupal\helfi_navigation\Service\GlobalNavigationService
     arguments: ['@cache.default', '@http_client', '@helfi_api_base.environment_resolver', '@language_manager', '@logger.channel.helfi_navigation', '@request_stack']
-  helfi_navigation.external_menu_block_base:
-    class: \Drupal\helfi_navigation\Plugin\Block\ExternalMenuBlockBase
-    arguments: ['@helfi_navigation.external_menu_tree_factory', '@helfi_navigation.global_navigation_service']
   helfi_navigation.menu_updater:
     class: \Drupal\helfi_navigation\MenuUpdater
-    arguments: ['@config.factory', '@entity_type.manager', '@helfi_navigation.global_navigation_service', '@language_manager', '@helfi_api_base.internal_domain_resolver']
+    arguments: ['@config.factory', '@helfi_navigation.global_navigation_service', '@language_manager', '@helfi_navigation.menu_tree_builder_service']
+  helfi_navigation.menu_tree_builder_service:
+    class: \Drupal\helfi_navigation\Service\MenuTreeBuilder
+    arguments: ['@config.factory', '@entity_type.manager', '@language_manager', '@helfi_api_base.internal_domain_resolver']
