+++ helfi_features/helfi_navigation/src/Plugin/Block/ExternalMenuBlock.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace Drupal\helfi_navigation\Plugin\Block;
+
+use Drupal\helfi_api_base\Environment\Project;
+
+/**
+ * Provides an external menu block.
+ *
+ * @Block(
+ *   id = "external_menu_block",
+ *   admin_label = @Translation("External menu block"),
+ *   category = @Translation("External menu")
+ * )
+ */
+class ExternalMenuBlock extends ExternalMenuBlockBase {
+
+  public function getData(): string {
+    // @todo Retrieve menu type from block settings.
+    $menu_type = 'main_menu';
+
+    return $this->globalNavigationService->makeRequest(
+      Project::ETUSIVU,
+      'GET',
+      "/global-menus/$menu_type"
+    );
+  }
+}
